Refactored packages. Updated urlToMobCapture to the correct 1.13.1 version. Updated version.
