+++ src/net/johnbrooks/mh/Commands.java
@@ -1,5 +1,6 @@
 package net.johnbrooks.mh;
 
+import net.johnbrooks.mh.managers.UpdateManager;
 import net.md_5.bungee.api.chat.ClickEvent;
 import net.md_5.bungee.api.chat.TextComponent;
 import org.bukkit.ChatColor;
