+++ src/plugin.yml
@@ -1,5 +1,5 @@
 name: MobCapture
-version: 1.1.12
+version: 1.2.0
 author: WiseHollow
 main: net.johnbrooks.mh.Main
 api-version: 1.13
